Change all data to unsigned ints
Change the float array to an unsigned int array. Change the mesh to
use glVertexAttribIPointer instead of glVertexAttribPointer when
taking in the vertex data. Change the shaders to take uvec's as
attribute types.

These changes were made to facilitate the process of combining all
vertex data into a single unsigned int. Instead of each vertex
using 5 floats, which is way too much memory, each vertex will
only use 1 unsigned int. Bitwise operators will be used to get
individual pieces of data from within that integer.

+++ res/shaders/basic_vertex.glsl
@@ -1,6 +1,6 @@
 #version 330 core
-layout(location = 0) in vec3 a_position;
-layout(location = 1) in vec2 a_texCoords;
+layout(location = 0) in uvec3 a_position;
+layout(location = 1) in uvec2 a_texCoords;
 
 out vec2 v_texCoords;
